+++ CHANGELOG.md
@@ -5,6 +5,7 @@ It is based on Tidepool Blip 1.27.
 ## 3.9.0-beta - 2026-03-12
 ### Added
 - YLP-3613 Memos on DBL settings in Devices view
+- YLP-3617 As a user, I can see the devices changes history
 
 ### Changed
 - YLP-3815 Display update timestamp in yourloops